Lesson 2 homeworks
