--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_2/Task_3/Book.java
@@ -1,5 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_2.Task_3;
-
-class Book {
-
-}
