--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_2/Task_5/BankAccount.java
@@ -1,9 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_2.Task_5;
-
-class BankAccount {
-
-    public static void main(String[] args) {
-
-    }
-
-}
