--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_2/Task_8/Hello.java
@@ -1,18 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_2.Task_8;
-
-import java.util.Scanner;
-
-class Hello {
-
-    public static void main(String[] args) {
-        System.out.println("Enter your name ");
-        Scanner scanner = new Scanner(System.in);
-        String name = scanner.nextLine();
-
-        System.out.println("Hello " + name);
-
-
-    }
-
-
-}
