--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_5/Task_18/SizeMatters.java
@@ -1,9 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_5.Task_18;
-
-class SizeMatters {
-
-    public static void main(String[] args) {
-        System.out.println("Sorry, I have an errors.");
-        System.out.println("Please fix me!");
-    }
-}
