--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_6/Task_23/AttentionPlease.java
@@ -1,9 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_6.Task_23;
-
-class AttentionPlease {
-
-    public static void main(String[] args) {
-        System.out.println("Code outside the method will not compile!");
-
-    }
-}
