--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_6/Task_24/NoSleepNoFocus.java
@@ -1,8 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_6.Task_24;
-
-class NoSleepNoFocus {
-    public static void main(String[] args) {
-        System.out.println("I wanna coffee");
-        System.out.println("So bad");
-    }
-}
