--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_7/Task_28/Task1.java
@@ -1,4 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_7.Task_28;
-
-class Task1 {
-}
