--- src/main/java/students/nikolaj_antonov/lesson_2/lesson_2/level_7/Task_29/SumTwoNumbers.java
@@ -1,18 +1,0 @@
-package students.nikolaj_antonov.lesson_2.lesson_2.level_7.Task_29;
-
-class SumTwoNumbers {
-
-    public static void main(String[] args) {
-
-        int numberOne = 10;
-        int numberTwo = 20;
-
-        System.out.println("Number 1 = " + numberOne);
-        System.out.println("Number 2 = " + numberTwo);
-
-        int sum = numberOne + numberTwo;
-
-        System.out.println("Sum = " + sum);
-    }
-
-}
