+++ src/main/java/students/nikolaj_antonov/lesson_2/level_2/Task_4/Phone.java
@@ -1,0 +1,5 @@
+package students.nikolaj_antonov.lesson_2.level_2.Task_4;
+
+class Phone {
+
+}
