+++ src/main/java/students/nikolaj_antonov/lesson_2/level_2/Task_5/BankAccount.java
@@ -1,0 +1,9 @@
+package students.nikolaj_antonov.lesson_2.level_2.Task_5;
+
+class BankAccount {
+
+    public static void main(String[] args) {
+
+    }
+
+}
