+++ src/main/java/students/nikolaj_antonov/lesson_2/level_2/Task_7/LoginService.java
@@ -1,0 +1,10 @@
+package students.nikolaj_antonov.lesson_2.level_2.Task_7;
+
+class LoginService {
+
+    public static void main(String[] args) {
+        System.out.println("You can login!");
+    }
+
+
+}
