+++ src/main/java/students/nikolaj_antonov/lesson_2/level_5/Task_15/Println.java
@@ -1,0 +1,16 @@
+package students.nikolaj_antonov.lesson_2.level_5.Task_15;
+
+class Println {
+    public static void main(String[] args) {
+
+        System.out.println("Hello, ");
+        System.out.println("I ");
+        System.out.println("am ");
+        System.out.println("a ");
+        System.out.println("student.");
+        System.out.print("Hello, ");
+        System.out.print("I am a student.");
+
+
+    }
+}
