+++ src/main/java/students/nikolaj_antonov/lesson_2/level_7/Task_28/BankAccount.java
@@ -1,0 +1,5 @@
+package students.nikolaj_antonov.lesson_2.level_7.Task_28;
+
+class BankAccount {
+
+}
