+++ src/main/java/students/nikolaj_antonov/lesson_2/level_7/Task_28/LoginService.java
@@ -1,0 +1,4 @@
+package students.nikolaj_antonov.lesson_2.level_7.Task_28;
+
+class LoginService {
+}
