+++ src/main/java/students/nikolaj_antonov/lesson_2/level_7/Task_30/CorrectDataTypes.java
@@ -1,0 +1,18 @@
+package students.nikolaj_antonov.lesson_2.level_7.Task_30;
+
+class CorrectDataTypes {
+
+    public static void main(String[] args) {
+
+        int numberOne = 10;
+        double numberTwo = 20.0;
+
+        System.out.println("Number 1 = " + numberOne);
+        System.out.println("Number 2 = " + numberTwo);
+
+        double sum = numberOne + numberTwo;
+
+        System.out.println("Sum = " + sum);
+    }
+
+}
